+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixes
 - General
   - not properly converting bit based sizes (KiB, MiB, etc...) to byte based
+  - uninstall and downgrade logs are not available [#255](https://github.com/vinifmor/bauh/issues/255)
   
 - Debian
   - info: crashing when the package size has unexpected symbols [#251](https://github.com/vinifmor/bauh/issues/251)